+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for zooming in and out via the context menu of the webapp viewer.
 - Support for passing arbitrary options to the IMU boxes of a BoxModel via the `imubox_options` property.
+- Support for setting custom application icons for webapps via the `icon` property of `qmt.Webapp`.
 ## Changed
 - Change minimum Python version from 3.7 to 3.8 (and minimum scipy version to 1.8.0).
 - Updated JavaScript dependencies of webapp library.